Исправлены замечания

+++ backend/foodgram/api/views.py
@@ -25,7 +25,7 @@ from .services import create_shopping_list
 from rest_framework.pagination import PageNumberPagination
 
 
-class CustomUserViewSet(UserViewSet):
+class UserViewSet(UserViewSet):
     permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get_serializer_class(self):
@@ -54,15 +54,15 @@ class RecipeViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         user = self.request.user
+        queryset = (Recipe.objects.select_related('author')
+                    .prefetch_related('ingredients'))
         if user.is_authenticated:
-            return (Recipe.objects.select_related('author')
-                    .prefetch_related('ingredients')
-                    .annotate(
-                        is_favorited=Exists(Favorite.objects.filter(
-                            user=user, recipe=OuterRef('pk'))),
-                        is_in_shopping_cart=Exists(ShoppingCart.objects.filter(
-                            user=user, recipe=OuterRef('pk')))))
-        return Recipe.objects.all()
+            return (queryset.annotate(
+                    is_favorited=Exists(Favorite.objects.filter(
+                        user=user, recipe=OuterRef('pk'))),
+                    is_in_shopping_cart=Exists(ShoppingCart.objects.filter(
+                        user=user, recipe=OuterRef('pk')))))
+        return queryset
 
     def perform_create(self, serializer):
         serializer.save(author=self.request.user)
@@ -71,10 +71,78 @@ class RecipeViewSet(viewsets.ModelViewSet):
         serializer.save(author=self.request.user)
         super(RecipeViewSet, self).perform_update(serializer)
 
+    @classmethod
+    def manage_user_recipe_relations(
+            cls, request, pk, queryset, related_field, serializer):
+        """
+        Обрабатывает просмотр рецептов,
+        добавление и удаление и связанных объектов).
+        """
+        if request.method == 'GET':
+            if pk is None:
+                queryset = queryset.filter(
+                    Q(**{f'{related_field}__user': request.user}))
+                serializer = RecipeGetSerializer(queryset, many=True)
+                return Response(serializer.data)
+            return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
+
+        if request.method == 'POST':
+            data = {
+                'user': request.user.id,
+                'recipe': pk,
+            }
+            serializer = serializer(data=data, context={'request': request})
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(RecipeShortSerializer(
+                get_object_or_404(Recipe, id=pk)).data,
+                status=status.HTTP_201_CREATED
+            )
+
+        if request.method == 'DELETE':
+            recipe = get_object_or_404(Recipe, id=pk)
+            del_count, _ = queryset.filter(Q(user=request.user)
+                                           & Q(recipe=recipe)).delete()
+            if del_count:
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
+
+    @action(['post', 'delete'], detail=True)
+    def favorite(self, request, pk=None):
+        if request.method == 'POST':
+            return self.manage_user_recipe_relations(
+                request, pk, queryset=Recipe.objects.all(),
+                related_field='favorited', serializer=FavoriteSerializer)
+        return self.manage_user_recipe_relations(
+            request, pk, queryset=Favorite.objects.all(),
+            related_field='favorited', serializer=FavoriteSerializer)
+
+    @action(['get'], detail=False)
+    def favorites(self, request, pk=None):
+        return self.manage_user_recipe_relations(
+            request, pk, queryset=Recipe.objects.all(),
+            related_field='favorited', serializer=FavoriteSerializer)
+
+    @action(['post', 'delete'], detail=True)
+    def shopping_cart(self, request, pk=None):
+        if request.method == 'POST':
+            return self.manage_user_recipe_relations(
+                request, pk, queryset=Recipe.objects.all(),
+                related_field='shopping', serializer=ShoppingCartSerializer)
+        return self.manage_user_recipe_relations(
+            request, pk, queryset=ShoppingCart.objects.all(),
+            related_field='shopping', serializer=ShoppingCartSerializer)
+
+    @action(['get'], detail=False, url_path='shopping_cart')
+    def get_shopping_cart(self, request, pk=None):
+        return self.manage_user_recipe_relations(
+            request, pk, queryset=Recipe.objects.all(),
+            related_field='shopping', serializer=ShoppingCartSerializer)
+
     @action(['get'], detail=False)
     def download_shopping_cart(self, request):
         temp_file = create_shopping_list(request.user)
-        temp_file.seek(0)
         response = FileResponse(temp_file, content_type='text/plain')
         response['Content-Disposition'] = (
             'attachment; '
@@ -123,90 +191,14 @@ class FollowView(APIView):
         serializer = FollowSerializer(
             data=data, context={'request': request}
         )
-        if serializer.is_valid():
-            serializer.save()
-            return Response(
-                UserSerializer(User.objects.get(pk=pk)).data,
-                status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(UserSerializer(User.objects.get(pk=pk)).data,
+                        status=status.HTTP_201_CREATED)
 
     def delete(self, request, pk):
         author = get_object_or_404(User, id=pk)
-        following = author.following.filter(user=request.user)
-        if following.exists():
-            following.delete()
+        del_count, _ = author.following.filter(user=request.user).delete()
+        if del_count:
             return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status.HTTP_400_BAD_REQUEST)
-
-
-class BaseRecipeView(APIView):
-    permission_classes = [IsAuthenticated]
-    serializer_class = RecipeGetSerializer
-
-    def function(self, request, pk, queryset, related_field, serializer):
-        if request.method == 'GET':
-            if pk is None:
-                queryset = queryset.filter(
-                    Q(**{f'{related_field}__user': request.user}))
-                serializer = RecipeGetSerializer(queryset, many=True)
-                return Response(serializer.data)
-            return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-        if request.method == 'POST':
-            data = {
-                'user': request.user.id,
-                'recipe': pk,
-            }
-            serializer = serializer(data=data, context={'request': request})
-            if serializer.is_valid():
-                serializer.save()
-                return Response(
-                    RecipeShortSerializer(
-                        get_object_or_404(Recipe, id=pk)).data,
-                    status=status.HTTP_201_CREATED
-                )
-            return Response(
-                serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-        if request.method == 'DELETE':
-            recipe = get_object_or_404(Recipe, id=pk)
-            added = queryset.filter(Q(user=request.user) & Q(recipe=recipe))
-            if added.exists():
-                added.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-            return Response(status.HTTP_400_BAD_REQUEST)
-        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-
-class FavoriteView(BaseRecipeView):
-    def get(self, request, pk=None):
-        return self.function(
-            request, pk, queryset=Recipe.objects.all(),
-            related_field='favorited', serializer=FavoriteSerializer)
-
-    def post(self, request, pk):
-        return self.function(
-            request, pk, queryset=Recipe.objects.all(),
-            related_field='favorited', serializer=FavoriteSerializer)
-
-    def delete(self, request, pk):
-        return self.function(
-            request, pk, queryset=Favorite.objects.all(),
-            related_field='favorited', serializer=FavoriteSerializer)
-
-
-class ShoppingCartView(BaseRecipeView):
-    def get(self, request, pk=None):
-        return self.function(
-            request, pk, queryset=Recipe.objects.all(),
-            related_field='shopping', serializer=ShoppingCartSerializer)
-
-    def post(self, request, pk):
-        return self.function(
-            request, pk, queryset=Recipe.objects.all(),
-            related_field='shopping', serializer=ShoppingCartSerializer)
-
-    def delete(self, request, pk):
-        return self.function(
-            request, pk, queryset=ShoppingCart.objects.all(),
-            related_field='shopping', serializer=ShoppingCartSerializer)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
